Refactor project structure and update WatchListButton component
- Moved WatchListButton and Profile components to a shared directory for better organization.
- Removed legacy Profile and WatchListButton components, consolidating functionality into shared components.
- Updated import paths across various files to reflect the new structure.
- Enhanced layout and responsiveness in page.tsx and related components.
- Cleaned up unused code and improved overall project maintainability.
