+++ app/page.tsx
@@ -2,7 +2,8 @@ import { ANIME } from "@consumet/extensions";
 import Image from "next/image";
 import Link from "next/link";
 import CarouselWithContent from "./_components/Carousel";
-import WatchListButton from "@/app/_components/WatchListButton/WatchListButton";
+import WatchListButton from "@/app/_components/shared/WatchListButton/WatchListButton";
+import AnimeCard from "./_components/AnimeCard";
 
 export const revalidate = 3600; // revalidate the data at most every hour
 
@@ -19,28 +20,9 @@ export default async function Home() {
         {/* Recent Episodes */}
         <aside className="w-full px-4">
           <h1 className="text-3xl font-semibold my-4">Recent Episodes</h1>
-          <div className="grid gap-4 grid-cols-[repeat(auto-fill,minmax(125px,1fr))]">
+          <div className="grid gap-2 grid-cols-[repeat(auto-fill,minmax(150px,1fr))]">
             {recentEpisodes.results.map((anime, index: number) => (
-              <div key={index}>
-                <Link
-                  // key={index}
-                  href={`anime/${anime.id}`}
-                  className="relative"
-                >
-                  <Image
-                    className="aspect-[3/4] object-cover rounded-md"
-                    src={anime.image as string}
-                    alt={anime.title as string}
-                    width={300}
-                    height={500}
-                    unoptimized
-                  />
-                  <h2 className="text-sm line-clamp-2 mt-1">
-                    {anime.title as string}
-                  </h2>
-                </Link>
-                <WatchListButton id={anime.id} title={anime.title as string} />
-              </div>
+              <AnimeCard anime={anime} key={index} />
             ))}
           </div>
         </aside>
@@ -49,9 +31,9 @@ export default async function Home() {
         <aside className="w-full pb-4 px-8 lg:pl-0 lg:pr-12 xl:max-w-[40vw]">
           <h1 className="font-semibold text-3xl my-4">Top Airing</h1>
           {topAiring.results.map((anime: any, index: number) => (
-            <Link className="flex my-4" key={index} href={`anime/${anime.id}`}>
+            <Link className="flex my-4 group/airing overflow-clip rounded-lg" key={index} href={`anime/${anime.id}`}>
               <Image
-                className="object-cover h-32 w-24 flex-none rounded-l-lg"
+                className="object-cover h-32 w-24 flex-none group-hover/airing:scale-110 transition-transform duration-200"
                 src={anime.image}
                 alt={anime.title}
                 width={300}
@@ -59,7 +41,7 @@ export default async function Home() {
                 unoptimized
               />
               {/* Anime Title and Genre */}
-              <div className="w-full p-2 bg-blue-gray-900 rounded-r-lg max-h-32 overflow-clip">
+              <div className="w-full p-2 bg-blue-gray-900 max-h-32 overflow-clip">
                 <h1 className="m-1 text-sm line-clamp-2">{anime.title}</h1>
                 <div className="flex gap-2 flex-wrap">
                   {anime.genres.map((genre: string, index: number) => (
