+++ app/watch-list/page.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { getServerSession } from "next-auth/next";
 import { authOptions } from "@/app/lib/authOptions";
-import WatchListButton from "@/app/_components/WatchListButton/WatchListButton";
+import WatchListButton from "@/app/_components/shared/WatchListButton/WatchListButton";
 
 export default async function WatchList() {
   const session = await getServerSession(authOptions as any);
